% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/brotli.R
\name{brotli}
\alias{brotli}
\alias{brotli_decode}
\alias{brotli_encode}
\title{Brotli Compression}
\usage{
brotli_encode(from, mode = c("generic", "text", "font"), quality = 11,
  log_win = 22, log_block = 0)

brotli_decode(from)
}
\description{
Compress and decompress binary data with Brotli.
}
\seealso{
memCompress
}

